backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
openai>=1.57.0
youtube-transcript-api==1.2.4
Pillow==11.0.0
pydantic==2.9.2
